api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["bld.ai Team <contact@bld.ai>"]

[tool.poetry.dependencies]
argon2-cffi = "^21.1.0"
boto3 = "^1.19.5"
celery = "^5.2.1"
dj-rest-auth = "^2.1.11"
Django = "3.2.15"
django-allauth = "^0.49.0"
django-cors-headers = "^3.10.0"
django-environ = "^0.8.1"
django-extensions = "^3.1.3"
django-filter = "^21.1"
django-fsm = "^2.7.1"
django-fsm-admin = "^1.2.4"
django-health-check = "^3.16.4"
django-model-utils = "^4.1.1"
django-redis = "^5.1.0"
django-rest-knox = "^4.1.0"
django-ses = "^2.3.0"
django-smuggler = "^1.0.3"
django-storages = "^1.11.1"
django-timezone-field = "<5.0"
django-versatileimagefield = "^2.2"
djangorestframework = "^3.12.4"
djangorestframework-camel-case = "^1.2.0"
drf-extra-fields = "^3.1.1"
drf-recaptcha = "2.0.7"
drf-spectacular = "^0.21.1"
drf-writable-nested = "^0.6.3"
edx-django-sites-extensions = "^4.0.0"
gunicorn = "^20.1.0"
Pillow = "^9.0.0"
psycopg2-binary = "^2.9.1"
python = "^3.9"
python-slugify = "^6.1.1"
pytz = "^2021.3"
requests = "^2.26.0"
uvicorn = "^0.17.5"
whitenoise = "^6.0.0"
ddtrace = "^0.60.0"
toml = "^0.10.2"
django-jsonform = "^2.9.0"
django-requestlogs = "^0.5.0"
django-ipware = "^4.0.2"
JSON-log-formatter = "^0.5.1"
django-modeltranslation = "^0.18.2"
django-csp = "^3.7"
web3 = "^5.30.0"
sentry-sdk = "^1.9.8"
enverus-developer-api = "^3.2.0"
django-resized = "^1.0.2"
django-celery-beat = "^2.2.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
django-coverage-plugin = "^2.0.0"
django-stubs = "^1.9.0"
djangorestframework-stubs = "^1.4.0"
factory-boy = "^3.2.0"
ipython = "^8.1.1"
isort = "^5.9.3"
mock = "^4.0.3"
mypy = "^0.931"
pip = "^21.2.4"
pylint = "^2.11.1"
pylint-django = "^2.4.4"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"
pytest-django = "^4.4.0"
pytest-dotenv = "^0.5.2"
pytest-env = "^0.6.2"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.6.1"
pytest-watch = "^4.2.0"
pytest-xdist = "^2.4.0"
safety = "^1.10.3"
django-silk = "^4.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
